Update Ollama model settings logic
