+++ src/js/settings/components/provider-settings/ollama.js
@@ -2,9 +2,12 @@
  * WordPress dependencies
  */
 import { useSelect, useDispatch } from '@wordpress/data';
-// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
-import { __experimentalInputControl as InputControl } from '@wordpress/components';
+import {
+	__experimentalInputControl as InputControl, // eslint-disable-line @wordpress/no-unsafe-wp-apis
+	SelectControl,
+} from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
+import { useEffect } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -35,8 +38,55 @@ export const OllamaSettings = ( { isConfigured = false } ) => {
 	const { setProviderSettings } = useDispatch( STORE_NAME );
 	const onChange = ( data ) => setProviderSettings( providerName, data );
 
+	const models = [
+		{ label: __( '-- Choose Model --', 'classifai' ), value: '' },
+	];
+
+	// Convert providerSettings.models to an array from an object.
+	if (
+		providerSettings?.models &&
+		! Array.isArray( providerSettings.models )
+	) {
+		for ( const [ key, value ] of Object.entries(
+			providerSettings.models
+		) ) {
+			models.push( { label: value, value: key } );
+		}
+	}
+
+	// Auto-select the first model if no model is selected and models are available.
+	useEffect( () => {
+		if (
+			! providerSettings?.model &&
+			models.length > 1 &&
+			models[ 1 ]?.value
+		) {
+			onChange( { model: models[ 1 ].value } );
+		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [ providerSettings?.models, providerSettings?.model ] );
+
+	// Check if we have actual models (not just the placeholder).
+	const hasModels = models.length > 1;
+
 	return (
 		<>
+			<SettingsRow
+				label={ __( 'Model', 'classifai' ) }
+				description={ __(
+					'Choose the model you want to use for requests. If no models are shown or you want to use a different model, please ensure this is installed in Ollama first.',
+					'classifai'
+				) }
+			>
+				<SelectControl
+					id={ `${ providerName }_model` }
+					onChange={ ( value ) => onChange( { model: value } ) }
+					value={ providerSettings?.model || '' }
+					options={ models }
+					disabled={ ! hasModels }
+					__nextHasNoMarginBottom
+				/>
+			</SettingsRow>
 			{ ! isConfigured && (
 				<OllamaBaseSettings
 					providerSettings={ providerSettings }
